+++ ui/packages/catalog-ui-search/src/main/webapp/react-component/menu-action/menu-action.tsx
@@ -14,7 +14,7 @@
  **/
 import * as React from 'react'
 import styled from 'styled-components'
-import MenuItem from '../presentation/menu-item'
+import MenuItem from '../menu-item'
 import { withDropdown, withContext, ContextType } from '../presentation/dropdown'
 import { hot } from 'react-hot-loader'
 import { Subtract } from '../../typescript'
