+++ ui/packages/catalog-ui-search/src/main/webapp/react-component/menu-item/index.tsx
